Unify util files
